Create calculator.rb
Sorry, I didn't think we were turning this project in so I experimented with methods. After using if and case to calculate the operations, I switched to a ruby built-in method that would substitute in the operation as a string and then carry out the calculation. Please let me know if there are any questions or if I should redo with if/case statements instead. Thanks!